--- cmd/bzip/bzip.go
@@ -1,37 +1,0 @@
-package main
-
-import (
-	"fmt"
-	"io/ioutil"
-	"log"
-
-	"github.com/yeka/zip"
-)
-
-func main() {
-
-	r, err := zip.OpenReader("encrypted.zip")
-	if err != nil {
-		log.Fatal(err)
-	}
-	defer r.Close()
-
-	for _, f := range r.File {
-		if f.IsEncrypted() {
-			f.SetPassword("12345")
-		}
-
-		r, err := f.Open()
-		if err != nil {
-			log.Fatal(err)
-		}
-
-		buf, err := ioutil.ReadAll(r)
-		if err != nil {
-			log.Fatal(err)
-		}
-		defer r.Close()
-
-		fmt.Printf("Size of %v: %v byte(s)\n", f.Name, len(buf))
-	}
-}
